+++ src/RcppExports.cpp
@@ -83,6 +83,30 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// dCov
+double dCov(NumericMatrix x, NumericMatrix y);
+RcppExport SEXP _MESS_dCov(SEXP xSEXP, SEXP ySEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type y(ySEXP);
+    rcpp_result_gen = Rcpp::wrap(dCov(x, y));
+    return rcpp_result_gen;
+END_RCPP
+}
+// dCor
+double dCor(NumericMatrix x, NumericMatrix y);
+RcppExport SEXP _MESS_dCor(SEXP xSEXP, SEXP ySEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type y(ySEXP);
+    rcpp_result_gen = Rcpp::wrap(dCor(x, y));
+    return rcpp_result_gen;
+END_RCPP
+}
 // filldown
 SEXP filldown(SEXP x);
 RcppExport SEXP _MESS_filldown(SEXP xSEXP) {
@@ -215,6 +239,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_MESS_cmd", (DL_FUNC) &_MESS_cmd, 2},
     {"_MESS_colCumSum", (DL_FUNC) &_MESS_colCumSum, 1},
     {"_MESS_cumsumbinning", (DL_FUNC) &_MESS_cumsumbinning, 4},
+    {"_MESS_dCov", (DL_FUNC) &_MESS_dCov, 2},
+    {"_MESS_dCor", (DL_FUNC) &_MESS_dCor, 2},
     {"_MESS_filldown", (DL_FUNC) &_MESS_filldown, 1},
     {"_MESS_hwe_frequencies", (DL_FUNC) &_MESS_hwe_frequencies, 3},
     {"_MESS_ks_cumtest", (DL_FUNC) &_MESS_ks_cumtest, 3},